res made the following changes: - Moved keycolor baking out of csImageManipulate into its own   class csBakeKeyColor. Also improved the color fill for   transparent areas a bit. - Fixed some warnings in csfgen. - A bit more cleanup and simplification of GL render-to-texture.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20886 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/csfgen/csfgen.cpp
@@ -417,7 +417,7 @@ static bool Convert (const char *fontfile)
 
   {
     int oldsize = opt.fontsize;
-    opt.fontsize = font->GetSize ();
+    opt.fontsize = (int)font->GetSize ();
     if (opt.fontsize != oldsize)
       csPrintf ("Could not set font size %d, using size %d\n",
         oldsize, opt.fontsize);
